extract UlightWasm into separate module
