+++ www/live_edit.html
@@ -15,6 +15,7 @@
     <script>0</script>
 
     <script src="/js/live_edit_core.js" type="module"></script>
+    <script src="/js/ulight.js" type="module"></script>
     <script src="/js/live_edit.js" type="module" async></script>
 
     <header>
